navbar & footer updated
