+++ src/styles/globals.css
@@ -81,4 +81,5 @@
     @apply text-sm;
     /* @apply 3xl:text-lg; */
   }
+
 }
